Updates for better code design
